Compiler: Make 'version' optional
This will be non-null when referenced from TestRun, but null when
referenced from TestResult. At least for now. Maybe I can figure out how
to track compiler versions and paths for the mutators/mutatees in the
test suite in the future. That would actually be a good piece of
information to have. It's not a responsibility here, so this change is
all we need.