TestRunResult: track more details about the run
is_single_step: The tests can be run in one of two modes: 'grouped' or
'single step'.

num_parallel_tests: The test driver can run multiple tests in parallel

num_omp_threads: This is actually a critical piece of information that
was just completely overlooked in the old dashboard. We also want to
eventually have thread scaling tests.

time: The elapsed time for a build (BuildResults.time) or an execution
of the test suite (TestRunResult.time). This will help us detect changes
in the amount of time required to build Dyninst or run the tests. This
has been a long-requested feature that I just never had time to
implement.